+++ src/repl/commands/events.rs
@@ -18,6 +18,10 @@ pub enum Setting {
     LineNumbers,
     /// System clipboard integration
     Clipboard,
+    /// Tab stop width
+    TabStop,
+    /// Expand tab setting (insert spaces instead of tab)
+    ExpandTab,
 }
 
 /// Values for settings
@@ -27,6 +31,8 @@ pub enum SettingValue {
     On,
     /// Disable the setting
     Off,
+    /// Numeric value for the setting
+    Number(usize),
 }
 
 /// Events that commands can produce to request changes